Config: Adjust bed size to 229x187 for the Petsfang print head.
The Petsfang is quite big in terms of required space on the bed. Since
we are using a glass bed with metal clamps, these clamps take some space
on the bed where the print head shouldn't crash with them. This reduces
the bed size to 229x187. If we need more space, we need another print
head.